chore: reorganizes analyzer tests to improve maintainability
